+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.2.0-dev
 
 - Enhancements
-    - Trying to quote invalid identifier types (such as functions) now raises an `ArgumentError`
+    - Trying to quote invalid identifier or value types (such as functions) now raises an `ArgumentError`
 
 ## v0.1.0 (2018-04-14)